Add SEO metadata and sitemap for quiz pages
