+++ public/robots.txt
@@ -3,7 +3,7 @@ User-agent: *
 Allow: /
 
 # Host
-Host: http:/quiz.theboringeducation.com/
+Host: https://quiz.theboringeducation.com/
 
 # Sitemaps
-Sitemap: http:/quiz.theboringeducation.com/sitemap.xml
+Sitemap: https://quiz.theboringeducation.com/sitemap.xml
